pyproject.toml
[project]
name = "image-ssl"
version = "0.1.0"
description = "Self-Supervised Learning techniques for images - exploring visual equivalents to next token prediction"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "numpy>=1.24.0",
    "matplotlib>=3.7.0",
    "pillow>=9.5.0",
    "tqdm>=4.65.0",
    "pyyaml>=6.0",
    "scikit-learn>=1.2.0",
    "opencv-python>=4.7.0",
    "scikit-image>=0.20.0",
    "scipy>=1.10.0",
    "networkx>=3.0",
]

[project.optional-dependencies]
tracking = [
    "tensorboard>=2.12.0",
    "wandb>=0.15.0",
]
dev = [
    "ruff>=0.8.0",
    "mypy>=1.0.0",
    "pytest>=7.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["techniques", "data", "models", "utils", "experiments"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # do not perform function calls in argument defaults
    "B905",  # zip without strict parameter
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # Allow unused imports in __init__.py

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false  # Set to true once we add type hints
ignore_missing_imports = true

[dependency-groups]
dev = [
    "mypy>=1.18.2",
    "pytest>=8.4.2",
    "ruff>=0.14.3",
]
